+++ boston/tensorflow/keras/README.md
@@ -117,24 +117,6 @@ gsutil cp -r data/boston_housing.npz gs://your-bucket-name/boston_housing.npz
 The `setup.py` file is run on the Cloud ML Engine server to install
 packages/dependencies and set a few options.
 
-```python
-from setuptools import find_packages
-from setuptools import setup
-
-with open('requirements.txt') as f:
-  requirements = [l.strip('\n') for l in f if
-                  l.strip('\n') and not l.startswith('#')]
-                  
-setup(
-  name='boston',
-  version='0.1',
-  install_requires=requirements,
-  packages=find_packages(),
-  include_package_data=True,
-  description='CMLE samples'
-)
-```
-
 Technically, Cloud ML Engine [requires a TensorFlow application to be
 pre-packaged] so that it can install it on the servers it spins up. However, if
 you supply a `setup.py` in the project root directory, then Cloud ML Engine will
@@ -192,7 +174,7 @@ rm -rf $JOB_DIR
 Run the model with python (local)
 
 ```
-python -m trainer.task --train_file=$TRAIN_FILE --job_dir=$JOB_DIR
+python -m trainer.task --train-file=$TRAIN_FILE --job-dir=$JOB_DIR
 ```
 
 ## Training using gcloud local
@@ -213,7 +195,7 @@ export TRAIN_FILE=gs://cloud-samples-data/ml-engine/boston/boston_housing.npz
 You can run Keras training using gcloud locally.
 
 ```
-gcloud ml-engine local train --module-name=trainer.task --package-path=trainer -- --train_file=$DATASET_FILE --job_dir=$JOB_DIR
+gcloud ml-engine local train --module-name=trainer.task --package-path=trainer -- --train-file=$TRAIN_FILE --job-dir=$JOB_DIR
 ```
 
 *Feel free to modify the destination file for in utils.py
@@ -236,7 +218,13 @@ export TRAIN_FILE=gs://cloud-samples-data/ml-engine/boston/boston_housing.npz
 You can train the model on Cloud ML Engine
 
 ```
-gcloud ml-engine jobs submit training $JOB_NAME --stream-logs --runtime-version 1.10 --job_dir=$JOB_DIR --package-path=trainer --module-name trainer.task --region $REGION -- --train_file=$DATASET_FILE
+gcloud ml-engine jobs submit training $JOB_NAME \
+    --stream-logs \
+    --runtime-version 1.10 \
+    --job-dir=$JOB_DIR \
+    --package-path=trainer \
+    --module-name trainer.task \
+    --region $REGION -- --train-file=$TRAIN_FILE
 ```
 
 ## Monitor training with TensorBoard
