+++ boston/tensorflow/keras/trainer/task.py
@@ -19,22 +19,15 @@ from __future__ import print_function
 import argparse
 import logging
 import os
-import subprocess
 import sys
 
 from . import model
+from . import utils
 
 import numpy as np
 import tensorflow as tf
 
 from tensorflow.contrib.training.python.training import hparam
-
-# Current working directory.
-WORKING_DIR = os.getcwd()
-# Temporary directory inside
-TEMP_DIR = 'tmp/'
-# Download file
-BOSTON_FILE = 'boston_housing.npz'
 
 
 def get_args():
@@ -45,40 +38,35 @@ def get_args():
 	"""
 	parser = argparse.ArgumentParser()
 	parser.add_argument(
-		'--job_dir',
+		'--job-dir',
 		type=str,
-		required=True,
 		help='GCS location to write checkpoints and export models')
 	parser.add_argument(
-		'--train_file',
+		'--train-file',
 		type=str,
 		required=True,
 		help='Dataset file local or GCS')
 	parser.add_argument(
-		'--test_split',
+		'--test-split',
 		type=float,
 		default=0.2,
 		help='Split between training and test, default=0.2')
 	parser.add_argument(
-		'--num_epochs',
+		'--num-epochs',
 		type=float,
 		default=500,
 		help='number of times to go through the data, default=500')
 	parser.add_argument(
-		'--batch_size',
+		'--batch-size',
 		type=int,
 		default=128,
 		help='number of records to read during each training step, default=128')
 	parser.add_argument(
-		'--learning_rate',
+		'--learning-rate',
 		type=float,
 		default=.001,
 		help='learning rate for gradient descent, default=.001')
-	parser.add_argument(
-		'--verbosity',
-		choices=['DEBUG', 'ERROR', 'FATAL', 'INFO', 'WARN'],
-		default='INFO',
-	)
+
 	return parser.parse_args()
 
 
@@ -103,92 +91,6 @@ def _setup_logging():
 	os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
 
-def download_files_from_gcs(source, destination):
-  """Download files from GCS to a WORKING_DIR/.
-
-  Args:
-    source: GCS path to the training data
-    destination: GCS path to the validation data.
-
-  Returns:
-    A list to the local data paths where the data is downloaded.
-  """
-  local_file_names = [destination]
-  gcs_input_paths = [source]
-
-  # Copy raw files from GCS into local path.
-  raw_local_files_data_paths = [os.path.join(WORKING_DIR, local_file_name)
-    for local_file_name in local_file_names
-    ]
-  for i, gcs_input_path in enumerate(gcs_input_paths):
-    if gcs_input_path:
-      subprocess.check_call(
-        ['gsutil', 'cp', gcs_input_path, raw_local_files_data_paths[i]])
-
-  return raw_local_files_data_paths
-
-
-
-def load_data(path='boston_housing.npz', test_split=0.2, seed=113):
-	"""Loads the Boston Housing dataset.
-
-	Args:
-		path: path where to cache the dataset locally (relative to
-			~/.keras/datasets).
-		test_split: fraction of the data to reserve as test set.
-		seed: Random seed for shuffling the data before computing the test split.
-
-	Returns:
-		Tuple of Numpy arrays: `(x_train, y_train), (x_test, y_test)`.
-
-	Raises:
-		ValueError: No dataset file defined.
-	"""
-	assert 0 <= test_split < 1
-	if not path:
-		raise ValueError('No dataset file defined')
-
-	if path.startswith('gs://'):
-		download_files_from_gcs(path, destination=BOSTON_FILE)
-		path = BOSTON_FILE
-
-	with np.load(path) as f:
-		x = f['x']
-		y = f['y']
-
-	np.random.seed(seed)
-	indices = np.arange(len(x))
-	np.random.shuffle(indices)
-	x = x[indices]
-	y = y[indices]
-
-	x_train = np.array(x[:int(len(x) * (1 - test_split))])
-	y_train = np.array(y[:int(len(x) * (1 - test_split))])
-	x_test = np.array(x[int(len(x) * (1 - test_split)):])
-	y_test = np.array(y[int(len(x) * (1 - test_split)):])
-	return (x_train, y_train), (x_test, y_test)
-
-
-def normalize_data(train_data, test_data):
-	"""Normalize features with different scales and ranges.
-
-	Subtract the mean of the feature and divide by the standard deviation.
-	Test data is *not* used when calculating the mean and std.
-
-	Args:
-		train_data: (numpy.darray) Training data.
-		test_data: (numpy.darray) Testing data.
-
-	Returns:
-		A tuple of training and test data.
-	"""
-	mean = train_data.mean(axis=0)
-	std = train_data.std(axis=0)
-	train_data = (train_data - mean) / std
-	test_data = (test_data - mean) / std
-	return train_data, test_data
-
-
 def train_and_evaluate(hparams):
 	"""Helper function: Trains and evaluate model.
 
@@ -198,7 +100,7 @@ def train_and_evaluate(hparams):
 	# Load data.
 	(train_data,
 	 train_labels), (test_data,
-									 test_labels) = load_data(path=hparams.train_file)
+									 test_labels) = utils.load_data(path=hparams.train_file)
 
 	# Shuffle data.
 	order = np.argsort(np.random.random(train_labels.shape))
@@ -206,15 +108,16 @@ def train_and_evaluate(hparams):
 	train_labels = train_labels[order]
 
 	# Normalize features with different scales and ranges.
-	train_data, test_data = normalize_data(train_data, test_data)
+	train_data, test_data = utils.normalize_data(train_data, test_data)
 
 	# Running configuration.
 	run_config = tf.estimator.RunConfig(save_checkpoints_steps=500)
 
-	# Create TrainSpec.
+	# Training steps
 	train_steps = hparams.num_epochs * len(train_data) / hparams.batch_size
 	# Reshape Label numpy array.
 	train_labels = np.asarray(train_labels).astype('float32').reshape((-1, 1))
+	# Create TrainSpec.
 	train_spec = tf.estimator.TrainSpec(
 		input_fn=lambda: model.input_fn(
 			train_data,
@@ -223,10 +126,11 @@ def train_and_evaluate(hparams):
 			mode=tf.estimator.ModeKeys.TRAIN),
 		max_steps=train_steps)
 
-	# Create EvalSpec.
+	# Create exporter information.
 	exporter = tf.estimator.LatestExporter('exporter', model.serving_input_fn)
 	# Reshape Label numpy array.
 	test_labels = np.asarray(test_labels).astype('float32').reshape((-1, 1))
+	# Create EvalSpec.
 	eval_spec = tf.estimator.EvalSpec(
 		input_fn=lambda: model.input_fn(
 			test_data,
